#46 Seguridad a las APIs de Valoration añadida Se ha definido "valuer" como "solo lectura" en ambos tipos de Valoration

+++ backend/apis/api_valoration/api.py
@@ -1,15 +1,21 @@
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes, authentication_classes
 from rest_framework import status
 
+from rest_framework.permissions import IsAuthenticated
+from rest_framework_simplejwt.authentication import JWTAuthentication
+from apis.permissions_decorators import IsStudentOrAdmin, IsAdmin
+from rest_framework.exceptions import PermissionDenied
 
-from apis.models import UserValoration, PropertyValoration
+from apis.models import UserValoration, PropertyValoration, Owner
 from .serializers import UserValorationSerializer, PropertyValorationSerializer
 
 ##################################################
 #Users
 
 @api_view(['GET', 'POST'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def valoration_user_api_view(request):
 
     if request.method == 'GET':
@@ -21,7 +27,13 @@ def valoration_user_api_view(request):
     
     #create
     elif request.method == 'POST':
+        
         data = request.data
+        data['valuer'] = request.user.id
+
+        if data['valuer'] == int(data['valued']):
+            return Response({'message':'No puede valorarse a si mismo'}, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = UserValorationSerializer(data = data) 
         if serializer.is_valid():
             serializer.save()
@@ -32,6 +44,8 @@ def valoration_user_api_view(request):
 
 
 @api_view(['GET', 'DELETE'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def valoration_user_detail_api_view(request, id):
         
     # queryset
@@ -46,6 +60,9 @@ def valoration_user_detail_api_view(request, id):
         
         
         elif request.method == 'DELETE':
+            if not (request.user.id == valoration.valuer.id or request.user.isAdministrator):
+                return Response({"message":"No puede borrar una valoración de otro usuario registrado"}, status=status.HTTP_403_FORBIDDEN)
+
             valoration.delete()
             return Response({'message':"Valoración a usuario eliminada correctamente!"}, status=status.HTTP_200_OK)
         
@@ -58,6 +75,8 @@ def valoration_user_detail_api_view(request, id):
 # Properties
 
 @api_view(['GET', 'POST'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def valoration_property_api_view(request):
 
     if request.method == 'GET':
@@ -69,7 +88,17 @@ def valoration_property_api_view(request):
     
     #create
     elif request.method == 'POST':
+       
         data = request.data
+        data['valuer'] = request.user.id
+        
+        if Owner.objects.filter(id=data['valuer']).exists():
+            return Response({'message':"Un propietario no puede valorar inmuebles"}, status=status.HTTP_403_FORBIDDEN)
+        
+        if request.user.isAdministrator:
+            return Response({'message':"Un administrador no puede valorar inmuebles"}, status=status.HTTP_403_FORBIDDEN)
+
+
         serializer = PropertyValorationSerializer(data = data) 
         if serializer.is_valid():
             serializer.save()
@@ -80,6 +109,8 @@ def valoration_property_api_view(request):
 
 
 @api_view(['GET', 'DELETE'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def valoration_property_detail_api_view(request, id):
         
     # queryset
@@ -94,6 +125,9 @@ def valoration_property_detail_api_view(request, id):
         
         
         elif request.method == 'DELETE':
+            if not (request.user.id == valoration.valuer.id or request.user.isAdministrator):
+                return Response({"message":"No puede borrar una valoración de otro estudiante"}, status=status.HTTP_403_FORBIDDEN)
+
             valoration.delete()
             return Response({'message':"Valoración a un inmueble eliminada correctamente!"}, status=status.HTTP_200_OK)
